+++ packages/node-utils/src/__tests__/extend.ts
@@ -1,6 +1,0 @@
-import { deepMerge } from '..'
-import extend from './extended'
-
-deepMerge({})
-
-export default extend