+++ packages/node-utils/src/__tests__/require.ts
@@ -1,7 +1,0 @@
-import { deepMerge } from '..'
-
-const required = require('./required')
-
-deepMerge({})
-
-export default required